// src/utils/types.ts
export type SafeNumber = number | `${number}`;

export type IProfileTabs =
  | "Student Profile"
  | "Study Preferances"
  | "Test Information"
  | "Assigned Agent"
  | "Transaction"
  | "Compaign Data";

export type IAgentTabs =
  | "Profile Details"
  | "Banking Details"
  | "Employee"
  | "Student";